feat : detach Home Modal from Home.tsx
